tactical_triage_env/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tactical_triage_env"
version = "1.0.0"
description = "Emergency 911 Dispatch Triage — OpenEnv RL Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "pydantic>=2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
]

[project.scripts]
server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["tactical_triage_env"]
